# server/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Registry API
  description: >
    Centralized registry for MCP servers and AI agents.
    Exposes a versioned HTTP API; all UIs are clients of this API.
  version: 0.4.0-rc0
  contact:
    name: AI Registry Team
servers:
  - url: /
    description: Current host
...
paths:
...
  /api/v1/groups:
    get:
      operationId: listGroups
      summary: List groups (Server Admin)
      tags: [rbac]
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/CursorParam"
      responses:
        "200":
          description: Paginated groups
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GroupList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createGroup
      summary: Create a group (Server Admin)
      tags: [rbac]
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateGroupRequest" }
      responses:
        "201":
          description: The created group
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Group" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /api/v1/groups/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getGroup
      summary: Get a group by slug (Server Admin)
      tags: [rbac]
      security: [{ cookieAuth: [] }]
      responses:
        "200":
          description: The group
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Group" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: patchGroup
      summary: Update a group (Server Admin)
      tags: [rbac]
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchGroupRequest" }
      responses:
        "200":
          description: The updated group
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Group" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteGroup
      summary: Delete a group (Server Admin)
      tags: [rbac]
      security: [{ cookieAuth: [] }]
      responses:
        "204":
          description: Deleted
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/groups/{slug}/members:
    get:
      operationId: listGroupMembers
      summary: List a group's members (Server Admin)
      tags: [rbac]
      security: [{ cookieAuth: [] }]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: The group's members
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/groups/{slug}/members/{email}:
    parameters:
      - name: slug
        in: path
        required: true
        schema: { type: string }
      - name: email
        in: path
        required: true
        schema: { type: string, format: email }
    put:
      operationId: addGroupMember
      summary: Add a member to a group by email (Server Admin)
      description: Auto-creates an invited user when the email is unknown.
      tags: [rbac]
      security: [{ cookieAuth: [] }]
      responses:
        "200":
          description: The added (or auto-created) user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: removeGroupMember
      summary: Remove a member from a group (Server Admin)
      tags: [rbac]
      security: [{ cookieAuth: [] }]
      responses:
        "204":
          description: Removed
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/users:
    get:
      operationId: listUsers
      summary: List users (Server Admin)
      tags: [rbac]
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/CursorParam"
      responses:
        "200":
          description: Paginated users
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createUser
      summary: Create a local or invited user (Server Admin)
      tags: [rbac]
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUserRequest" }
      responses:
        "201":
          description: The created user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: ai_registry_session
      description: >
        Registry session cookie set by the login endpoints. Secure + HttpOnly;
        the browser sends it automatically. There is no bearer token.
  parameters:
...
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CursorParam:
      name: cursor
      in: query
      schema:
        type: string
  schemas:
...
    Group:
      type: object
      required: [id, slug, name, created_at, updated_at]
      properties:
        id: { type: string }
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    GroupList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Group" }
        next_cursor: { type: string }
    CreateGroupRequest:
      type: object
      required: [slug, name]
      properties:
        slug: { type: string }
        name: { type: string }
        description: { type: string }
    PatchGroupRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
    User:
      type: object
      required: [id, email, has_password, is_server_admin, disabled, created_at, updated_at]
      properties:
        id: { type: string }
        email: { type: string, format: email }
        display_name: { type: string }
        subject:
          type: string
          description: OIDC subject; empty for local-only users.
        has_password: { type: boolean }
        is_server_admin: { type: boolean }
        disabled: { type: boolean }
        last_seen_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    UserList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/User" }
        next_cursor: { type: string }
    CreateUserRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
        display_name: { type: string }
        password:
          type: string
          format: password
          description: Optional; omit to create an invited user that cannot log in until a credential is set.
        is_server_admin: { type: boolean }
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  responses:
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: Authenticated but lacks required scope
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: Duplicate slug or version
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ValidationError:
      description: Request body failed schema validation
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...